[IMP] website_rating: lightly lint portal chatter
PURPOSE

Allow publishers to add a comment on frontend reviews with ratings. Purpose
is to allow to give some feedback or context on published reviews.

SPECIFICATIONS

  * rename / reorder some methods according to JS guidelines;
  * split portal composer and portal chatter override to better understand
    module organization and ease future code additions;

LINKS

Task ID : 2026911
