+++ addons/website_rating/views/assets.xml
@@ -6,6 +6,7 @@
         </xpath>
         <xpath expr="//script[last()]" position="after">
             <script type="text/javascript" src="/website_rating/static/src/js/portal_chatter.js"></script>
+            <script type="text/javascript" src="/website_rating/static/src/js/portal_composer.js"></script>
             <script type="text/javascript" src="/website_rating/static/src/js/portal_rating_composer.js"></script>
         </xpath>
     </template>
